requirements.txt
Django>=5.0,<6.0
gunicorn
openpyxl>=3.1.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
whitenoise>=6.5.0
dj-database-url>=2.1.0
psycopg2-binary>=2.9.9
